Add new bipartite graph class and Matches class, to replace the Correspondences class. The bipartite graph reimplementation is based on a sketch from Pau Gargallo, which uses the better map of maps structure to store the graph.
The Matches class has a better iterator API for accessing features that are in
an image, or features that are part of a given track. It also supports
filtering by feature type, so that it is easy, for example, to draw all point
features (while the drawing code is not aware that the point features are
actually SURFPointFeatures with a descriptor and scale).

This will soon replace the existing correspondences*.cc/h and bipartite_graph*.